Excel反序列成POJO, 阶段性完成, 接下来考虑为@Col注解添加converter属性, JSR303 Validator注解的支持

+++ loser-workbook/src/test/java/com/loserico/workbook/pojo/SettlementItem.java
@@ -45,7 +45,7 @@ public class SettlementItem {
 	private String remark; // 备注
 
 	@Col(name = "SKU编码")
-	private Long skuCode; // SKU编码
+	private Integer skuCode; // SKU编码
 
 	@Col(name = "SKU名称")
 	private String skuName; // SKU名称
